# swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "description":"This is api for restaurants",
        "version":"__version__"
    },
    "tags":[
        {
        "name":"PUBLIC ROUTE",
        "description":"Public Route"
        }
    ],
    "paths":{
        "/health":{
            "summary":"Testing Health Check",
            "get":{
                "description":"Testing User for app",
                "tags":["PUBLIC ROUTE"],
                "responses":{
                    "200":{
                        "description":"Health Check Pass"
                    }
                }
            }
        },
        "/location":{
            "summary":"Testing location route",
            "get":{
                "description":"Testing location route for app",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"x-access-auth",
                        "in":"header",
                        "required":true
                    }
                ],
                "responses":{
                    "200":{
                        "description":"location route"
                    }
                }
            }
        },
        "/details":{
            "summary":"Testing location route",
            "get":{
                "description":"Testing location route for app",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "required":true
                    },
                    {
                        "name":"mealid",
                        "in":"query"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"location route"
                    }
                }
            }
        },
        "/placeOrder":{
            "summary":"Testing location route",
            "post":{
                "description":"Testing location route for app",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "name":"Khushi",
                                "email":"Khushi@gmail.com",
                                "address":"Hom 65",
                                "phone":83333211,
                                "cost":533,
                                "menuItem":[
                                    9,4,5
                                ],
                                "status":"pending"
                            }
                        }
                    }
                }
            }
        },
        "/updateOrder":{
            "summary":"Testing location route",
            "put":{
                "description":"Testing location route for app",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "id":"7999999000",
                                "status":"pending"
                            }
                        }
                    }
                },
               "responses":{
                    "200":{
                        "description":"location route"
                    }
                }
            }
        },
        "/deleteOrder":{
            "summary":"Testing location route",
            "delete":{
                "description":"Testing location route for app",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "id":"7999999000",
                                "status":"pending"
                            }
                        }
                    }
                },
               "responses":{
                    "200":{
                        "description":"location route"
                    }
                }
            }
        }
        
    }
}